Object-Oriented Programming in JS
I learned how to put the functionality in an object that can ever be used per user.
I also learned that you can separate a method, an object's function, on its own object to save on my computer's memory resources. 
When we do that separation, we look up the object's prototype to find that object's function.